// mobile/types/server/ImageMessageEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MessageReply } from "./MessageReply";

export type ImageMessageEvent = { image_data: string, filename: string, mime_type: string, size: number, width: number | null, height: number | null, reply_to: MessageReply | null, };